Fix:금액/비율 타입 BigDecimal로 통일

+++ core-api/src/main/java/com/pocketstock/core/asset/dto/CategoryAmountRow.java
@@ -2,8 +2,10 @@ package com.pocketstock.core.asset.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CategoryAmountRow {
     private String category;
-    private long amount;
+    private BigDecimal amount;
 }
